feat(x86_64): sync router live configs, optimize plugins and add compression to CI

+++ build_helper/prepare.py
@@ -102,8 +102,6 @@ def prepare(configs: dict[str, dict[str, Any]]) -> None:
     logger.info("开始克隆拓展软件源码...")
     to_clone: set[tuple[str, str]] = {("https://github.com/immortalwrt/packages", ""),
                                        ("https://github.com/chenmozhijin/turboacc", "package"),
-                                       ("https://github.com/pymumu/openwrt-smartdns", "master"),
-                                       ("https://github.com/pymumu/luci-app-smartdns", "master"),
                                        *[(pkg["REPOSITORIE"], pkg["BRANCH"]) for config in configs.values() for pkg in config["extpackages"].values()],
                                        *[("https://github.com/sbwml/packages_lang_golang",
                                           config["openwrtext"]["golang_version"]) for config in configs.values()]}
@@ -164,39 +162,40 @@ def prepare(configs: dict[str, dict[str, Any]]) -> None:
             shutil.copytree(os.path.join(openwrt_paths, cfg_names[0]), os.path.join(openwrt_paths, name), symlinks=True)
     openwrts = {name: OpenWrt(os.path.join(openwrt_paths, name), configs[name]["compile"]["openwrt_tag/branch"]) for name in cfg_names}
 
-    # 下载AdGuardHome规则与配置
-    logger.info("下载AdGuardHome规则与配置...")
     global_files_path = os.path.join(paths.workdir, "files")
     shutil.copytree(os.path.join(paths.openwrt_k, "files"), global_files_path, symlinks=True)
-    adg_filters_path = os.path.join(global_files_path, "usr", "bin", "AdGuardHome", "data", "filters")
-    os.makedirs(adg_filters_path, exist_ok=True)
-    filters = {"1628750870.txt": "https://adguardteam.github.io/AdGuardSDNSFilter/Filters/filter.txt",
-               "1628750871.txt": "https://anti-ad.net/easylist.txt",
-               "1677875715.txt": "https://easylist-downloads.adblockplus.org/easylist.txt",
-               "1677875716.txt": "https://easylist-downloads.adblockplus.org/easylistchina.txt",
-               "1677875717.txt": "https://raw.githubusercontent.com/cjx82630/cjxlist/master/cjx-annoyance.txt",
-               "1677875718.txt": "https://raw.githubusercontent.com/zsakvo/AdGuard-Custom-Rule/master/rule/zhihu-strict.txt",
-               "1677875720.txt": "https://gist.githubusercontent.com/Ewpratten/a25ae63a7200c02c850fede2f32453cf/raw/b9318009399b99e822515d388b8458557d828c37/hosts-yt-ads",
-               "1677875724.txt": "https://raw.githubusercontent.com/banbendalao/ADgk/master/ADgk.txt",
-               "1677875725.txt": "https://www.i-dont-care-about-cookies.eu/abp/",
-               "1677875726.txt": "https://raw.githubusercontent.com/jdlingyu/ad-wars/master/hosts",
-               "1677875727.txt": "https://raw.githubusercontent.com/Goooler/1024_hosts/master/hosts",
-               "1677875728.txt": "https://winhelp2002.mvps.org/hosts.txt",
-               "1677875733.txt": "https://raw.githubusercontent.com/hl2guide/Filterlist-for-AdGuard/master/filter_whitelist.txt",
-               "1677875734.txt": "https://raw.githubusercontent.com/hg1978/AdGuard-Home-Whitelist/master/whitelist.txt",
-               "1677875735.txt": "https://raw.githubusercontent.com/mmotti/adguard-home-filters/master/whitelist.txt",
-               "1677875737.txt": "https://raw.githubusercontent.com/liwenjie119/adg-rules/master/white.txt",
-               "1677875739.txt": "https://raw.githubusercontent.com/JamesDamp/AdGuard-Home---Personal-Whitelist/master/AdGuardHome-Whitelist.txt",
-               #"1677875740.txt": "https://raw.githubusercontent.com/scarletbane/AdGuard-Home-Whitelist/main/whitelist.txt"
-    }
-    dl_tasks: list[DLTask] = []
-    for name, url in filters.items():
-        dl_tasks.append(dl2(url, os.path.join(adg_filters_path, name)))
 
-    dl_tasks.append(dl2("https://raw.githubusercontent.com/chenmozhijin/AdGuardHome-Rules/main/AdGuardHome-dnslist(by%20cmzj).yaml",
-                     os.path.join(global_files_path, "etc", "AdGuardHome-dnslist(by cmzj).yaml")))
+    # 如果有配置开启了 AdGuardHome 则下载规则与配置
+    if any("CONFIG_PACKAGE_luci-app-adguardhome=y" in config.get("openwrt", "") for config in configs.values()):
+        logger.info("下载AdGuardHome规则与配置...")
+        adg_filters_path = os.path.join(global_files_path, "usr", "bin", "AdGuardHome", "data", "filters")
+        os.makedirs(adg_filters_path, exist_ok=True)
+        filters = {"1628750870.txt": "https://adguardteam.github.io/AdGuardSDNSFilter/Filters/filter.txt",
+                   "1628750871.txt": "https://anti-ad.net/easylist.txt",
+                   "1677875715.txt": "https://easylist-downloads.adblockplus.org/easylist.txt",
+                   "1677875716.txt": "https://easylist-downloads.adblockplus.org/easylistchina.txt",
+                   "1677875717.txt": "https://raw.githubusercontent.com/cjx82630/cjxlist/master/cjx-annoyance.txt",
+                   "1677875718.txt": "https://raw.githubusercontent.com/zsakvo/AdGuard-Custom-Rule/master/rule/zhihu-strict.txt",
+                   "1677875720.txt": "https://gist.githubusercontent.com/Ewpratten/a25ae63a7200c02c850fede2f32453cf/raw/b9318009399b99e822515d388b8458557d828c37/hosts-yt-ads",
+                   "1677875724.txt": "https://raw.githubusercontent.com/banbendalao/ADgk/master/ADgk.txt",
+                   "1677875725.txt": "https://www.i-dont-care-about-cookies.eu/abp/",
+                   "1677875726.txt": "https://raw.githubusercontent.com/jdlingyu/ad-wars/master/hosts",
+                   "1677875727.txt": "https://raw.githubusercontent.com/Goooler/1024_hosts/master/hosts",
+                   "1677875728.txt": "https://winhelp2002.mvps.org/hosts.txt",
+                   "1677875733.txt": "https://raw.githubusercontent.com/hl2guide/Filterlist-for-AdGuard/master/filter_whitelist.txt",
+                   "1677875734.txt": "https://raw.githubusercontent.com/hg1978/AdGuard-Home-Whitelist/master/whitelist.txt",
+                   "1677875735.txt": "https://raw.githubusercontent.com/mmotti/adguard-home-filters/master/whitelist.txt",
+                   "1677875737.txt": "https://raw.githubusercontent.com/liwenjie119/adg-rules/master/white.txt",
+                   "1677875739.txt": "https://raw.githubusercontent.com/JamesDamp/AdGuard-Home---Personal-Whitelist/master/AdGuardHome-Whitelist.txt",
+        }
+        dl_tasks: list[DLTask] = []
+        for name, url in filters.items():
+            dl_tasks.append(dl2(url, os.path.join(adg_filters_path, name)))
 
-    wait_dl_tasks(dl_tasks)
+        dl_tasks.append(dl2("https://raw.githubusercontent.com/chenmozhijin/AdGuardHome-Rules/main/AdGuardHome-dnslist(by%20cmzj).yaml",
+                         os.path.join(global_files_path, "etc", "AdGuardHome-dnslist(by cmzj).yaml")))
+
+        wait_dl_tasks(dl_tasks)
 
     # 获取用户信息
     logger.info("编译者：%s", compiler)
@@ -245,13 +244,14 @@ def prepare_cfg(config: dict[str, Any],
     shutil.copytree(os.path.join(cloned_repos[("https://github.com/immortalwrt/packages", "")], "admin", "netdata"),
                         os.path.join(openwrt.path, "feeds", "packages", "admin", "netdata"), symlinks=True)
     # 更新smartdns
-    shutil.rmtree(os.path.join(openwrt.path, "feeds", "luci", "applications", "luci-app-smartdns"), ignore_errors=True)
-    shutil.rmtree(os.path.join(openwrt.path, "feeds", "packages", "net", "smartdns"), ignore_errors=True)
-    shutil.copytree(cloned_repos[("https://github.com/pymumu/luci-app-smartdns", "master")],
-                    os.path.join(openwrt.path, "feeds", "luci", "applications", "luci-app-smartdns"), symlinks=True)
-    shutil.copytree(cloned_repos[("https://github.com/pymumu/openwrt-smartdns", "master")],
-                    os.path.join(openwrt.path, "feeds", "packages", "net", "smartdns"), symlinks=True)
-    disable_smartdns_hash_check(os.path.join(openwrt.path, "feeds", "packages", "net", "smartdns", "Makefile"))
+    if ("https://github.com/pymumu/luci-app-smartdns", "master") in cloned_repos:
+        shutil.rmtree(os.path.join(openwrt.path, "feeds", "luci", "applications", "luci-app-smartdns"), ignore_errors=True)
+        shutil.rmtree(os.path.join(openwrt.path, "feeds", "packages", "net", "smartdns"), ignore_errors=True)
+        shutil.copytree(cloned_repos[("https://github.com/pymumu/luci-app-smartdns", "master")],
+                        os.path.join(openwrt.path, "feeds", "luci", "applications", "luci-app-smartdns"), symlinks=True)
+        shutil.copytree(cloned_repos[("https://github.com/pymumu/openwrt-smartdns", "master")],
+                        os.path.join(openwrt.path, "feeds", "packages", "net", "smartdns"), symlinks=True)
+        disable_smartdns_hash_check(os.path.join(openwrt.path, "feeds", "packages", "net", "smartdns", "Makefile"))
 
     logger.info("%s处理软件包...", cfg_name)
     for pkg_name, pkg in config["extpackages"].items():
